Include a fader inbetween brightness increments
Now since the brightness increment calculation occurs every minute, the gap between the two values when updated is a somewhat noticable but not too bad. Decided to throw in a fader() to blend the brightness to make the project feel more complete. Although I am not including an outfader. The jolt seems to catch my eye which is actually better when to glows to purple/red.

Adjust Minute Sleep timer. Should be 7.5 times for a minute, but I think because of the reduced 1KHz clock, it runs a bit slower? So eight 8 second sleep cycles covers close to 58 seconds on my board.