Upload STLs
A future idea if I were to do it again, would be to use a magnetic switch(Hall Effect Sensor), that way I dont need to precisely measure the height of the button/cap/holder & its switch distance and create the top cover to account for all of those while keeping the differ wall hidden, while also trying to maintain a low profile. Another reason is because the switch and the top cover without any filler/foam will introduce an uneven balance point because of the switch being in contact with both the base and top cover, so I had to use some foam to make it balanced. A magnetic sensor will allow the two pieces slide past each other without being in physical contact.