Add glow on button press
because of that one glitch on the digispark where if the button is pressed too quickly and then the chip going to sleep right after, causes some unstability somewhere, adding this glow inditator upon placing a cup both gave a indiator that the chip is not stuck in some state, as well as seemingly to reduce that glitch as well probably by acting as a psedo debouncer too since the switch is now stable.

also it seems the mega has the button to be instant interrupts on pin 21.
Or its possibly internal damaged because i dont have a resistor even though I have a internal pull up resistor? Seems to only work on LOW trigger now and doesnt do anything on CHANGE, even on a quick test program i made. so idk.